+++ packages/editor/editor-common/src/providerFactory/index.ts
@@ -25,7 +25,7 @@ export default class ProviderFactory {
 
   setProvider(name: string, provider?: Promise<any>) {
     // Do not trigger notifyUpdate if provider is the same.
-    if (provider && this.providers.get(name) === provider) {
+    if (this.providers.get(name) === provider) {
       return;
     }