+++ packages/media/media-test-helpers/src/mediaPickerMocks.ts
@@ -1,8 +1,14 @@
-import { State } from '../domain';
+import { State } from '@atlaskit/media-picker/src/popup/domain';
 import { Store } from 'react-redux';
 import { Observable } from 'rxjs/Observable';
 import { ContextFactory } from '@atlaskit/media-core';
 
+if (typeof jest === 'undefined') {
+  // We need to do this since jest is not defined on browser integration tests
+  (global as any).jest = {
+    fn: () => ({ mockReturnValue() {} }),
+  };
+}
 export const mockState: State = {
   redirectUrl: 'some-redirect-url',
   view: {
@@ -93,7 +99,7 @@ export const mockFetcher = () => ({
 });
 
 export const mockIsWebGLNotAvailable = () => {
-  jest.mock('../tools/webgl', () => {
+  jest.mock('@atlaskit/media-picker/src/popup/tools/webgl', () => {
     return {
       isWebGLAvailable: jest.fn(() => {
         return false;