+++ packages/people-and-teams/focused-task-close-account/src/i18n/es.ts
@@ -78,6 +78,40 @@ export default {
     'Es posible que tú u otros usuarios hayáis instalado aplicaciones que añadan funciones a los productos de Atlassian. Estas aplicaciones pueden haber almacenado la información del perfil del usuario.',
   'focused-task-close-account.delete-account.overview.inline-dialog.data-apps.self':
     'Es posible que tú u otros usuarios hayáis instalado aplicaciones que añadan funciones a los productos de Atlassian. Estas aplicaciones pueden haber almacenado la información de tu perfil.',
+  'focused-task-close-account.deactivate-account.overview.heading.self':
+    'Desactivar la cuenta',
+  'focused-task-close-account.deactivate-account.overview.heading.admin':
+    'Desactivar la cuenta',
+  'focused-task-close-account.deactivate-account.overview.first.line.self':
+    'Estás a punto de desactivar la cuenta de:',
+  'focused-task-close-account.deactivate-account.overview.first.line.admin':
+    'Estás a punto de desactivar la cuenta de:',
+  'focused-task-close-account.deactivate-account.overview.last.line.self':
+    'Puedes volver a activar la cuenta en cualquier momento.',
+  'focused-task-close-account.deactivate-account.overview.last.line.admin':
+    'Puedes volver a activar la cuenta en cualquier momento.',
+  'focused-task-close-account.deactivate-account.overview.paragraph.about-to-deactivate.admin':
+    'Cuando desactives la cuenta:',
+  'focused-task-close-account.deactivate-account.overview.paragraph.about-to-deactivate.self':
+    'Cuando desactives la cuenta:',
+  'focused-task-close-account.deactivate-account.overview.paragraph.loseAccess.admin':
+    '{fullName} <b>perderá el acceso de inmediato</b> a todos los servicios de la cuenta de Atlassian. En este momento, tiene acceso a:',
+  'focused-task-close-account.deactivate-account.overview.paragraph.loseAccess.self':
+    '<b>Perderás el acceso de inmediato</b> a todos los servicios de la cuenta de Atlassian. En este momento, tienes acceso a:',
+  'focused-task-close-account.deactivate-account.overview.paragraph.loseAccess.admin.noSites':
+    '{fullName} <b>perderá el acceso de inmediato</b> a todos los servicios de la cuenta de Atlassian. En este momento, no tiene acceso a ninguno de ellos, excepto a servicios como Community o Marketplace.',
+  'focused-task-close-account.deactivate-account.overview.paragraph.loseAccess.self.noSites':
+    '<b>Perderás el acceso de inmediato</b> a todos los servicios de la cuenta de Atlassian. En este momento, no tienes acceso a ninguno de ellos, excepto a servicios como Community y Marketplace.',
+  'focused-task-close-account.deactivate-account.overview.paragraph.loseAccess.footnote':
+    'Otros servicios de la cuenta de Atlassian, como Atlassian Community y Marketplace.',
+  'focused-task-close-account.deactivate-account.overview.paragraph.personal-data.admin':
+    'Sus datos personales, como su nombre completo y su dirección de correo electrónico, y el contenido que hayan creado permanecerán en los servicios de la cuenta de Atlassian.',
+  'focused-task-close-account.deactivate-account.overview.paragraph.personal-data.self':
+    'Sus datos personales, como su nombre completo y su dirección de correo electrónico, y el contenido que hayan creado permanecerán en los servicios de la cuenta de Atlassian.',
+  'focused-task-close-account.deactivate-account.overview.paragraph.billing.admin':
+    'Dejaremos de facturarte por estos usuarios.',
+  'focused-task-close-account.deactivate-account.overview.paragraph.billing.self':
+    'Dejaremos de facturarte por estos usuarios.',
   'focused-task-close-account.delete-account.content-preview.heading':
     'Cómo verán los usuarios esta cuenta',
   'focused-task-close-account.delete-account.drop-down-expand-button':